cuma_cloud/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic[email]>=2.5.0
pydantic-settings>=2.0.0
sqlalchemy[asyncio]>=2.0.23
asyncpg>=0.29.0
alembic>=1.13.0
python-dotenv>=1.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0
python-multipart>=0.0.6
